Book-Club-backend/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.2
click==8.1.8
distlib==0.3.9
Django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
fastapi==0.109.1
filelock==3.16.1
h11==0.16.0
idna==3.10
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.20
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.35.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.27.0
virtualenv==20.27.1
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
tqdm==4.66.1
faker==21.0.0
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
requests-mock==1.11.0
factory-boy==3.3.0

# Code quality and linting tools
black==23.7.0
isort==5.12.0
flake8==6.0.0
bandit==1.7.5
safety==2.3.4
mypy==1.4.1
django-stubs==4.2.3
types-requests==2.31.0

# Pre-commit hooks
pre-commit==3.3.3
